refactor: fe - 폴더구조 수정 #13
