--- frontend/src/components/workspace-list/index.style.tsx
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-
-export const Title = styled.p`
-	font-size: 32px;
-	font-weight: 700;
-	line-height: 44px;
-
-	margin-bottom: 28px;
-`;
