+++ frontend/src/data/workspace-sidebar.tsx
@@ -1,6 +1,5 @@
-import React from 'react';
-import RecentWorkspace from '@pages/main/recent-workspace.component';
-import AllWorkspace from '@pages/main/all-workspace.component';
+import RecentWorkspace from '@pages/main/recent-workspace';
+import AllWorkspace from '@pages/main/all-workspace';
 
 export const sidebarItems: SidebarItemType = {
 	recent: { id: 1, title: 'recents', component: <RecentWorkspace /> },
