+++ frontend/src/pages/main/contents/index.style.tsx
