+++ frontend/src/pages/main/index.tsx
@@ -1,8 +1,7 @@
-import React from 'react';
-import Contents from './contents.component';
-import Header from './header.component';
+import Contents from './contents';
+import Header from './header';
 import { Wrapper } from './index.style';
-import Sidebar from './sidebar.component';
+import Sidebar from './sidebar';
 
 function Main() {
 	return (
