+++ frontend/src/pages/main/recent-workspace/index.tsx
@@ -1,5 +1,4 @@
-import React from 'react';
-import WorkspaceList from '@components/workspace-list';
+import WorkspaceList from '@pages/main/workspace-list';
 
 function RecentWorkspace() {
 	return (
