+++ frontend/src/pages/main/sidebar/index.style.tsx
