+++ frontend/src/pages/main/workspace-card/index.style.tsx
