+++ frontend/src/pages/main/workspace-list/index.style.tsx
@@ -1,8 +1,16 @@
 import styled from 'styled-components';
 
+export const Title = styled.h1`
+	font-size: 32px;
+	font-weight: 700;
+	line-height: 44px;
+
+	margin-bottom: 28px;
+`;
+
 export const WorkspaceListContainer = styled.div`
 	display: flex;
-	gap: 1rem;
+	gap: 32px;
 	flex-basis: 150px;
 	flex-grow: 0;
 	flex-wrap: wrap;
