+++ frontend/src/pages/main/workspace-template-list/index.style.tsx
