+++ frontend/src/pages/workspace-template/index.style.tsx
